+++ src/Controllers/MenuFXMLController.java
@@ -9,9 +9,7 @@ import javafx.fxml.Initializable;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
-import javafx.event.ActionEvent;
-import javafx.fxml.FXML;
-import javafx.scene.text.Text;
+
 
 /**
  *
@@ -27,24 +25,6 @@ public class MenuFXMLController extends Controller implements Initializable{
      */
     public MenuFXMLController(){
         curView = "Menu";        
-    }
-    
-    /*
-     * FXML Scene Login
-     */
-    @FXML private Text actiontarget;
-    
-    @FXML protected void handleSubmitButtonAction(ActionEvent event) {
-        actiontarget.setText("Signing In");
-    }
-    public void start(Stage stage) throws Exception {
-       Parent root = FXMLLoader.load(getClass().getResource("MenuFXML.fxml"));
-    
-        Scene scene = new Scene(root, 300, 275);
-    
-        stage.setTitle("FXML Welcome");
-        stage.setScene(scene);
-        stage.show();
     }
